move other options as well
