+++ torch/csrc/api/include/torch/nn/modules/conv.h
@@ -2,6 +2,7 @@
 
 #include <torch/expanding_array.h>
 #include <torch/nn/cloneable.h>
+#include <torch/nn/options/conv.h>
 #include <torch/nn/pimpl.h>
 #include <torch/types.h>
 
@@ -12,69 +13,6 @@
 
 namespace torch {
 namespace nn {
-
-/// Options for a `D`-dimensional convolution module.
-template <size_t D>
-struct ConvOptions {
-  ConvOptions(
-      int64_t input_channels,
-      int64_t output_channels,
-      ExpandingArray<D> kernel_size) :
-                input_channels_(input_channels),
-                output_channels_(output_channels),
-                kernel_size_(std::move(kernel_size)) {}
-
-  /// The number of channels the input volumes will have.
-  /// Changing this parameter after construction __has no effect__.
-  TORCH_ARG(int64_t, input_channels);
-
-  /// The number of output channels the convolution should produce.
-  /// Changing this parameter after construction __has no effect__.
-  TORCH_ARG(int64_t, output_channels);
-
-  /// The kernel size to use.
-  /// For a `D`-dim convolution, must be a single number or a list of `D`
-  /// numbers.
-  /// This parameter __can__ be changed after construction.
-  TORCH_ARG(ExpandingArray<D>, kernel_size);
-
-  /// The stride of the convolution.
-  /// For a `D`-dim convolution, must be a single number or a list of `D`
-  /// numbers.
-  /// This parameter __can__ be changed after construction.
-  TORCH_ARG(ExpandingArray<D>, stride) = 1;
-
-  /// The padding to add to the input volumes.
-  /// For a `D`-dim convolution, must be a single number or a list of `D`
-  /// numbers.
-  /// This parameter __can__ be changed after construction.
-  TORCH_ARG(ExpandingArray<D>, padding) = 0;
-
-  /// The kernel dilation.
-  /// For a `D`-dim convolution, must be a single number or a list of `D`
-  /// numbers.
-  /// This parameter __can__ be changed after construction.
-  TORCH_ARG(ExpandingArray<D>, dilation) = 1;
-
-  /// For transpose convolutions, the padding to add to output volumes.
-  /// For a `D`-dim convolution, must be a single number or a list of `D`
-  /// numbers.
-  /// This parameter __can__ be changed after construction.
-  TORCH_ARG(ExpandingArray<D>, output_padding) = 0;
-
-  /// If true, convolutions will be transpose convolutions (a.k.a.
-  /// deconvolutions).
-  /// Changing this parameter after construction __has no effect__.
-  TORCH_ARG(bool, transposed) = false;
-
-  /// Whether to add a bias after individual applications of the kernel.
-  /// Changing this parameter after construction __has no effect__.
-  TORCH_ARG(bool, with_bias) = true;
-
-  /// The number of convolution groups.
-  /// This parameter __can__ be changed after construction.
-  TORCH_ARG(int64_t, groups) = 1;
-};
 
 /// Base class for all (dimension-specialized) convolution modules.
 template <size_t D, typename Derived>
@@ -114,9 +52,6 @@ class TORCH_API Conv1dImpl : public ConvImpl<1, Conv1dImpl> {
   Tensor forward(const Tensor& input);
 };
 
-/// `ConvOptions` specialized for 1-D convolution.
-using Conv1dOptions = ConvOptions<1>;
-
 /// A `ModuleHolder` subclass for `Conv1dImpl`.
 /// See the documentation for `Conv1dImpl` class to learn what methods it
 /// provides, or the documentation for `ModuleHolder` to learn about PyTorch's
@@ -134,9 +69,6 @@ class TORCH_API Conv2dImpl : public ConvImpl<2, Conv2dImpl> {
   Tensor forward(const Tensor& input);
 };
 
-/// `ConvOptions` specialized for 2-D convolution.
-using Conv2dOptions = ConvOptions<2>;
-
 /// A `ModuleHolder` subclass for `Conv2dImpl`.
 /// See the documentation for `Conv2dImpl` class to learn what methods it
 /// provides, or the documentation for `ModuleHolder` to learn about PyTorch's
@@ -153,9 +85,6 @@ class TORCH_API Conv3dImpl : public ConvImpl<3, Conv3dImpl> {
   using ConvImpl<3, Conv3dImpl>::ConvImpl;
   Tensor forward(const Tensor& input);
 };
-
-/// `ConvOptions` specialized for 3-D convolution.
-using Conv3dOptions = ConvOptions<3>;
 
 /// A `ModuleHolder` subclass for `Conv3dImpl`.
 /// See the documentation for `Conv3dImpl` class to learn what methods it
