+++ torch/csrc/api/include/torch/nn/modules/linear.h
@@ -2,6 +2,7 @@
 
 #include <torch/nn/cloneable.h>
 #include <torch/nn/module.h>
+#include <torch/nn/options/linear.h>
 #include <torch/nn/pimpl.h>
 #include <torch/types.h>
 
@@ -10,16 +11,6 @@
 
 namespace torch {
 namespace nn {
-/// Options for the `Linear` module.
-struct TORCH_API LinearOptions {
-  LinearOptions(int64_t in, int64_t out);
-  /// The number of input features (columns of the input matrix).
-  TORCH_ARG(int64_t, in);
-  /// The number of output features to produce (columns of the output matrix).
-  TORCH_ARG(int64_t, out);
-  /// Whether to learn and add a bias after the linear transformation.
-  TORCH_ARG(bool, with_bias) = true;
-};
 
 /// Applies a linear transformation with optional bias.
 class TORCH_API LinearImpl : public Cloneable<LinearImpl> {
