+++ torch/csrc/api/include/torch/nn/modules/rnn.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <torch/nn/cloneable.h>
+#include <torch/nn/options/rnn.h>
 #include <torch/nn/modules/dropout.h>
 #include <torch/nn/pimpl.h>
 #include <torch/types.h>
@@ -27,30 +28,6 @@ struct TORCH_API RNNOutput {
 };
 
 namespace detail {
-
-/// Common options for LSTM and GRU modules.
-struct TORCH_API RNNOptionsBase {
-  RNNOptionsBase(int64_t input_size, int64_t hidden_size);
-  virtual ~RNNOptionsBase() = default;
-  /// The number of features of a single sample in the input sequence `x`.
-  TORCH_ARG(int64_t, input_size);
-  /// The number of features in the hidden state `h`.
-  TORCH_ARG(int64_t, hidden_size);
-  /// The number of recurrent layers (cells) to use.
-  TORCH_ARG(int64_t, layers) = 1;
-  /// Whether a bias term should be added to all linear operations.
-  TORCH_ARG(bool, with_bias) = true;
-  /// If non-zero, adds dropout with the given probability to the output of each
-  /// RNN layer, except the final layer.
-  TORCH_ARG(double, dropout) = 0.0;
-  /// Whether to make the RNN bidirectional.
-  TORCH_ARG(bool, bidirectional) = false;
-  /// If true, the input sequence should be provided as `(batch, sequence,
-  /// features)`. If false (default), the expected layout is `(sequence, batch,
-  /// features)`.
-  TORCH_ARG(bool, batch_first) = false;
-};
-
 /// Base class for all RNN implementations (intended for code sharing).
 template <typename Derived>
 class TORCH_API RNNImplBase : public torch::nn::Cloneable<Derived> {
@@ -139,38 +116,6 @@ class TORCH_API RNNImplBase : public torch::nn::Cloneable<Derived> {
 
 // ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ RNN ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-enum class RNNActivation : uint32_t {ReLU, Tanh};
-
-/// Options for RNN modules.
-struct TORCH_API RNNOptions {
-  RNNOptions(int64_t input_size, int64_t hidden_size);
-
-  /// Sets the activation after linear operations to `tanh`.
-  RNNOptions& tanh();
-  /// Sets the activation after linear operations to `relu`.
-  RNNOptions& relu();
-
-  /// The number of features of a single sample in the input sequence `x`.
-  TORCH_ARG(int64_t, input_size);
-  /// The number of features in the hidden state `h`.
-  TORCH_ARG(int64_t, hidden_size);
-  /// The number of recurrent layers (cells) to use.
-  TORCH_ARG(int64_t, layers) = 1;
-  /// Whether a bias term should be added to all linear operations.
-  TORCH_ARG(bool, with_bias) = true;
-  /// If non-zero, adds dropout with the given probability to the output of each
-  /// RNN layer, except the final layer.
-  TORCH_ARG(double, dropout) = 0.0;
-  /// Whether to make the RNN bidirectional.
-  TORCH_ARG(bool, bidirectional) = false;
-  /// If true, the input sequence should be provided as `(batch, sequence,
-  /// features)`. If false (default), the expected layout is `(sequence, batch,
-  /// features)`.
-  TORCH_ARG(bool, batch_first) = false;
-  /// The activation to use after linear operations.
-  TORCH_ARG(RNNActivation, activation) = RNNActivation::ReLU;
-};
-
 /// A multi-layer Elman RNN module with Tanh or ReLU activation.
 /// See https://pytorch.org/docs/master/nn.html#torch.nn.RNN to learn about the
 /// exact behavior of this module.
@@ -200,8 +145,6 @@ TORCH_MODULE(RNN);
 
 // ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ LSTM ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-using LSTMOptions = detail::RNNOptionsBase;
-
 /// A multi-layer long-short-term-memory (LSTM) module.
 /// See https://pytorch.org/docs/master/nn.html#torch.nn.LSTM to learn about the
 /// exact behavior of this module.
@@ -225,8 +168,6 @@ class TORCH_API LSTMImpl : public detail::RNNImplBase<LSTMImpl> {
 TORCH_MODULE(LSTM);
 
 // ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ GRU ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-using GRUOptions = detail::RNNOptionsBase;
 
 /// A multi-layer gated recurrent unit (GRU) module.
 /// See https://pytorch.org/docs/master/nn.html#torch.nn.GRU to learn about the
