+++ torch/csrc/api/include/torch/nn/options/loss.h
@@ -1,0 +1,18 @@
+#pragma once
+
+#include <torch/arg.h>
+
+namespace torch {
+namespace nn {
+
+/// Options for a L1 loss module.
+struct TORCH_API L1LossOptions {
+  L1LossOptions(Reduction::Reduction reduction = Reduction::Mean)
+      : reduction_(reduction) {}
+
+  /// Specifies the reduction to apply to the output.
+  TORCH_ARG(Reduction::Reduction, reduction);
+};
+
+} // namespace nn
+} // namespace torch
