+++ torch/csrc/api/src/nn/options/pooling.cpp
@@ -1,0 +1,11 @@
+#include <torch/nn/options/pooling.h>
+
+namespace torch {
+namespace nn {
+
+template struct MaxPoolOptions<1>;
+template struct MaxPoolOptions<2>;
+template struct MaxPoolOptions<3>;
+
+} // namespace nn
+} // namespace torch
